♻️ Rename index.ts to db.ts
